Complete all plugin requirements and code standards fixes
Extract library functions:
- Move local_grpcalendarimport_parse_csv() and local_grpcalendarimport_create_event()
  from index.php to locallib.php for unit testing
- Update index.php to require locallib.php and remove duplicates

Fix PSR2 multi-line call style:
- Extract $params variable before $DB->record_exists_sql() call
- Eliminates ContentAfterOpenBracket and MultipleArguments violations

Privacy provider:
- Implement null_provider for classes/privacy/provider.php
- No MOODLE_INTERNAL guard needed for autoloaded class files
- Add privacy:metadata language string

PHPUnit tests:
- Create final class in local_grpcalendarimport namespace
- Test parse_csv: basic CSV, TSV, BOM stripping, short-row padding, invalid path
- Test create_event: missing fields, invalid course/group, success, duplicate skip/not-skip

CI locale fix:
- Add locale-gen en_AU.UTF-8 step before PHPUnit
